+++ lemur/plugins/lemur_vault_dest/plugin.py
@@ -62,8 +62,8 @@ class VaultSourcePlugin(SourcePlugin):
             "name": "tokenFileOrVaultRole",
             "type": "str",
             "required": True,
-            "validation": check_validation("^([a-zA-Z0-9/._-]+/?)+$"),
-            "helpMessage": "Must be vaild file path for token based auth and valid role if k8s based auth",
+            "validation": check_validation("^[a-zA-Z0-9/._-]+/?$"),
+            "helpMessage": "Must be valid file path for token based auth and valid role if k8s based auth",
         },
         {
             "name": "vaultMount",
@@ -189,7 +189,7 @@ class VaultDestinationPlugin(DestinationPlugin):
             "name": "tokenFileOrVaultRole",
             "type": "str",
             "required": True,
-            "validation": check_validation("^([a-zA-Z0-9/._-]+/?)+$"),
+            "validation": check_validation("^[a-zA-Z0-9/._-]+/?$"),
             "helpMessage": "Must be vaild file path for token based auth and valid role if k8s based auth",
         },
         {
@@ -203,15 +203,17 @@ class VaultDestinationPlugin(DestinationPlugin):
             "name": "vaultPath",
             "type": "str",
             "required": True,
-            "validation": check_validation("^(([a-zA-Z0-9._-]+|{(CN|OU|O|L|S|C)})+/?)+$"),
-            "helpMessage": "Must be a valid Vault secrets path. Support vars: {CN|OU|O|L|S|C}",
+            "validation": check_validation(
+                "^([a-zA-Z0-9._-]+|{CN}|{OU}|{O}|{L}|{S}|{C})(/?([a-zA-Z0-9._-]+|{CN}|{OU}|{O}|{L}|{S}|{C}))*$"),
+            "helpMessage": "Must be a valid Vault secrets path. Support vars: {CN}|{OU}|{O}|{L}|{S}|{C}",
         },
         {
             "name": "objectName",
             "type": "str",
             "required": False,
-            "validation": check_validation("^([0-9a-zA-Z.:_-]+|{(CN|OU|O|L|S|C)})+$"),
-            "helpMessage": "Name to bundle certs under, if blank use {CN}. Support vars: {CN|OU|O|L|S|C}",
+            "validation": check_validation(
+                "^([a-zA-Z0-9:._-]+|{CN}|{OU}|{O}|{L}|{S}|{C})(/?([a-zA-Z0-9._-]+|{CN}|{OU}|{O}|{L}|{S}|{C}))*$"),
+            "helpMessage": "Name to bundle certs under, if blank use {CN}. Support vars: {CN}|{OU}|{O}|{L}|{S}|{C}",
         },
         {
             "name": "bundleChain",